pyproject.toml
[project]
name = "epstein-search"
version = "0.1.0"
description = "Detective document search system with OCR, semantic search, and RAG"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ollama",              # Official Ollama Python client
    "pymupdf",             # PDF handling (no external deps)
    "pillow",              # Image handling
    "chromadb",            # Vector store
    "rich",                # CLI output formatting
    "click",               # CLI framework
    "fastapi",             # Web framework
    "uvicorn[standard]",   # ASGI server
    "jinja2",              # HTML templating
    "python-multipart",    # Form handling
]

[project.scripts]
epstein-search = "main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src", "main.py"]
